:recycle: #852 - Some refactors.
