+++ frontend/zac-ui/libs/features/search/src/lib/search-form/object-search-form/zaak-object-search-form.component.html
@@ -1,6 +1,6 @@
 <gu-form [form]="form" buttonLabel="Zoeken" (formSubmit)="submitForm($event)"></gu-form>
 <mat-action-list>
   <mat-list-item *ngFor="let zaakObject of zaakObjects" (click)="selectZaakObject.emit(zaakObject)">
-    {{zaakObjectService.stringifyZaakObject(zaakObject)}}
+    {{ zaakObject | zaakObjectString }}
   </mat-list-item>
 </mat-action-list>
