+++ frontend/zac-ui/libs/features/search/src/lib/search-form/object-search-form/zaak-object.ts
