+++ frontend/zac-ui/libs/features/search/src/lib/search-form/search-form.component.html
@@ -12,7 +12,7 @@
       <!-- Case (zaak) by object object. -->
       <gu-tab label="Zoeken op zaakobject">
         <div class="mt-4">
-          <gu-object-search-form (selectZaakObject)="selectZaakObject($event)"></gu-object-search-form>
+          <gu-zaak-object-search-form (selectZaakObject)="selectZaakObject($event)"></gu-zaak-object-search-form>
         </div>
       </gu-tab>
     </gu-tab-group>
